### 2017-08-01 #334 @pjotrp
@cokelaer thank you for your submission! I'll be looking for a reviewer soon. In the mean time can you help out with a few things? I can't find the paper.md and paper.bib files. Can you point them to me? Also can you make the LICENSE file standard - if it is it will show up on github with the right license icon. Thanks!
