### 2017-08-02 #334 @cokelaer
@ pjotrp
I've fixed the LICENSE and the icon now appears on github.
As for the paper.md and paper.bib, what do mean exactly ? Where shall I provide those files and what should be the content ? thanks
