### 2017-08-02 #334 @pjotrp
Best is to check one of the existing publication. See https://github.com/genenetwork/genenetwork2/tree/master/doc/joss/2016
