### 2017-08-02 #334 @cokelaer
Do the paper.bib and paper.md must be in the repository of the software or can they be put elsewhere ? Sequana is an organisation and we put the software in sequana/sequana but for papers, we'd like to keep it separate for instance in ./sequana/resources . Would it be fine with your policy ?
