### 2017-08-02 #334 @arfon
> Would it be fine with your policy?

I *think* this is going to break some of our submission tooling (i.e. it's expected that the paper be in the software repository being reviewed). Let me check and get back to you.

For reference, some authors have made a folder in `docs` e.g..`docs/paper/paper.md` - might something like this work for you?
