### 2017-08-03 #334 @pjotrp
As the final paper gets generated from the source tree it is tied to the commit of that release. You *can* remove the paper from the source tree once it is published. It will be in that particular source release and git history. So it is not intrusive. Makes sense to tie it to the project, right?
