### 2017-08-03 #334 @cokelaer
okay, that's fine, I will add a ./joss directory in  the ./doc
thanks
