### 2017-08-03 #334 @cokelaer
FYI we have added a paper.bib paper.md in https://github.com/sequana/sequana/tree/master/doc/joss
