### 2017-08-03 #334 @pjotrp
That looks great. @bow since since this is a snakemake application would you care to review this JOSS submission?
