### 2017-08-04 sequana/sequana#430 @cokelaer
sequanix improvments
- [x] remove the sequana_logger_debug file ?
- [ ] mechanism to load pipelines from the web or a registry system ?
- [ ] tab with the cluster_config editable like the config file.
- [ ] Right now, we can use the mode "Input directory" or "input sample". In input sample, we can select 1 or 2 files only; would be nice to be able to select N samples
