### 2017-08-04 #352 @whedon
[REVIEW]: Sequana: a set of flexible genomic pipelines for processing and reporting NGS analysis
**Submitting author:** @cokelaer (<a href="http://orcid.org/0000-0001-6286-1138">Thomas Cokelaer</a>)
**Repository:** <a href="https://github.com/sequana/sequana/" target ="_blank">https://github.com/sequana/sequana/</a>
**Version:** 0.4.1
