### 2017-08-08 #352 @bow
I've reviewed the submission (the `paper.md`), the documentation, and I've also tried `sequana`, both as a user and as a developer. I will be opening a series of issues on the sequana repository soon. Before that, here are my general remarks.

This is a Python package that provides a slew of handy features both for pipeline developers and users alike. I could see this being used as a tool for Snakemake pipeline developers to collaborate with users who are less familiar with Snakemake in general, as mentioned in the submission.

The authors should be commended for the thorough documentation. From the online documentation in readthedocs.io, the docstrings in the source code, comments in the pipeline config files, and the tooltips in the Qt app, one could see that the authors have alloted significant, well-spent time for this task. Not to say that there are no room for improvements, of course. There were still some issues with the installation of dependencies and running the test examples. These will be addressed in the linked issues.

One concern I have is the monolithic structure of the package. As the first page of the documentation notes, sequana is now provides several distinct functionalities that could be packaged separately: a Python library for NGS analysis, a set of Snakemake pipelines, tools for creating the pipelines, and some standalone applications. Developers may only be interested in a subset of these (e.g. the library or the pipelines only), and providing them as a separate packages may encourage use.

This is not a blocking concern, however, and I think the submission should be accepted after the linked issues are addressed.

My final general remark is about the distinction of local vs cluster run in sequanix. It may be useful to add support controlling and/or monitoring a cloud run, given that Snakemake v4.0 supports running on Kubernetes. Additionally, it is now only possible to load Snakefiles and config files on a locally-mounted filesystem. It may also be useful to allow for users to input a URL pointing to these files and have sequanix download them. Combinining these two features would allow sequanix to be used as a native GUI app that can run public Snakefiles whose input files are also remotely stored.
