### 2017-08-08 sequana/sequana#436 @bow
JOSS review: developer setup
Hi @cokelaer,

After spending some time with sequana and trying to have a local environment setup, I found some things on the dependency handling that I believe can be improved.

Putting on a developer's hat, I was quite unsure about how to have a local development environment set up for developing sequana (not just the pipelines). While there is the `requirements.txt` file and the conda `environment.yml` file, I wasn't sure which one to use since they do not have exactly the same dependency. I tried looking into `setup.py` and `.travis.yml`, and in those places I also find dependencies that are different.

My suggestions are as follows:
1. Sync the overlapping dependencies in `requirements.txt`, `environment.yml`, and `.travis.yml`. For example, by adding a version qualifier for `easydev` in `requirements.txt` (since it exists in `environment.yml` as well).
2. If possible, include `requirements.txt` into `setup.py` instead of hard-coding the `install_requires` parameter.
3. Add the test dependencies in a separate file (e.g. `requirements-dev.txt`). This can then be included in the documentation (instead of hard-coding the dependencies [here](http://sequana.readthedocs.io/en/master/developers.html?highlight=pytest#testing-with-pytest)) and/or the `setup.py` file as value for the `test_require` parameter.
4. Update the documentation in the [developer guide](http://sequana.readthedocs.io/en/master/developers.html) section with instructions on setting up a local dev environment.

Additionally, I did stumble a bit when trying to run some of the included pipelines and tools. Have you looked into specifying a conda environment for each of the pipeline? Or maybe to use the Snakemake `conda` directive so that your pipelines are always guaranteed to pull their dependencies?

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
