### 2017-08-08 sequana/sequana#437 @bow
JOSS review: readthedocs.io overview section
I find that after running `conda install sequana` and running the example [here](http://sequana.readthedocs.io/en/master/userguide.html#sequana-library), I could not reproduce the plot. The standalone example with `sequana_coverage` works fine, though.

1. Is there something that needs to be setup that I may be missing?
2. Have you looked into using an jupyter notebook for this part? Maybe linking into an externally hosted notebook file, with the commands shown, is easier for the user?

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
