### 2017-08-08 sequana/sequana#438 @bow
JOSS review: libselinux dependency for sequanix
After `conda install sequana`, I could not get `sequanix` to immediately run. I was greeted instead with this message:

```
...
ImportError: libselinux.so.1: cannot open shared object file: No such file or directory
```

It turns out that `libselinux` is required by PyQt and it is not available in conda. This was solved by me installing `libselinux` from my distro package manager and then sequanix worked fine. I think it would be helpful to have this mentioned in the documentation, that libselinux is required and users may need to install it separately.

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
