### 2017-08-08 sequana/sequana#439 @bow
JOSS review: remove local hardcoded variables from pipeline
Upon checking the pipelines, I see that `markduplicates` in the `atac-seq` and `rna-seq` pipelines still contains a hard-coded command line flag: https://github.com/sequana/sequana/blob/master/sequana/pipelines/atac-seq/config.yaml#L57 and https://github.com/sequana/sequana/blob/4943ab9dd229a9a3548975fe7be67231f60115e2/sequana/pipelines/rnaseq/config.yaml.

Could these be removed?

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
