### 2017-08-08 sequana/sequana#440 @bow
JOSS review: typos
These are some collections to typos that I found during the review. I will reply to this thread if I find more. For now here are the list:

* `Since there are part of the Sequana` [docs, section 2.4](https://sequana.readthedocs.io/en/master/userguide.html#sequana-reports)
* `... of New Generation Sequencing (NGS) ` -> I believe it should be `Next Generation ...` [submission paper.md]
* `... which ease the decomposition of pipelines into modular sub-units.` [submission paper.md], should be `... which eases the decomposition of pipelines into modular sub-units.`
* `The reports are used to store the results of a pipelines but also materials required to reprodude the results.` [submission paper.md](https://github.com/sequana/sequana/blob/master/doc/joss/paper.md)
* `analysis tasks (e.g., variant calling filtering)` [submission paper.md], perhaps you meant `(e.g. variant calling)` or `(e.g. variant call filtering)`?
* ` as well as tool to perform` [submission paper.md], should be `as well as a tool to perform`
* ` test suite (on Travis.org) ` [submission paper.md]. The address is travis-ci.org, and this can be made into a hyperlink.
* ` and faster by taking care of the dependencies (e.g., samtools, bwa standalones, or Python librairies).` [submission paper.md], I think `bwa` (without `standalones` is sufficient) and it should be `libraries`.

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
