### 2017-08-08 sequana/sequana#441 @bow
JOSS review: tag release
This is more of a placeholder issue for after the review is done: a version tag (0.4.1) should be added to the git repository.

As a side note, it may also be useful to add git tags for previously released versions of sequana. This would show up in the [`releases` section](https://github.com/sequana/sequana/releases) nicely.

(Main review thread: [here](https://github.com/openjournals/joss-reviews/issues/352))
