### 2017-08-09 sequana/sequana#436 @cokelaer
Hi @bow ,
Great, that's very good points. We were discussing this issue recently and these suggestions are very useful. We will update the code soon to make the installation simpler for everybody and uniformise the setuptools / reathedocs / travis installation. Keep you posted once this is done.
