### 2017-08-09 #352 @pjotrp
Thanks @bow for the thorough review. @cokelaer can you take a look?
