### 2017-08-09 sequana/sequana#437 @cokelaer
strange, I just re-did the plot either from the link you posted or using the standalone.

1. the genome coverage code does not rely on specific bioinformatics package. It is a home-made tool based on numpy/pandas/matplotlib only so nothing special to set up
2. we have a notebook https://github.com/sequana/sequana/blob/master/notebooks/coverage.ipynb I have added a *see also* in the documentation (not yet built) that should appear on the develop branch later.

- [ ] check the notebook link
- [ ]  I will try the *sequana* version from conda to check whether I can reproduce the script and run the notebook
