### 2017-08-09 sequana/sequana#438 @cokelaer
- [ ] I will update the FAQs installation section with this issue and check whether we can fix this issue upstream
