### 2017-08-09 sequana/sequana#439 @cokelaer
good point, we will see how to update the pipeline so that it is fully generic. thanks again for pointing this.
We are reviewing the pipelines and rules those days with more documentation so we will check this point carefully.
