### 2017-08-09 sequana/sequana#440 @cokelaer
We fixed the typos mentioned here abive. thanks
