### 2017-08-09 sequana/sequana#441 @cokelaer
play, good point. We post releases on pypi in general but not on github but we could do both indeed.
