### 2017-08-09 #352 @cokelaer
hi, we are dealing with the issues right now (thanks again to @bow for the review and useful comments). One of the most important http://github.com/sequana/sequana#436 will be dealt with asap since this is relevant for the installation process. Other issues should be fixed very soon (or already fixed for those such as typos and documentation).

As for the general remarks here above, the two main ones concern
- monolithic structure: yes, this is true. We started with the aim of providing a couple of pipelines in snakemake format and somehow it has grown much bigger with the additions of pipelines, the graphical interface and original standalone applications. The idea to split Sequana into smaller sub-units could be envisage especially for Sequanix. This is not a small work though so we may do that on the mid-term but probably not in the next two months.
- Improvements of Sequanix with respect to new Snakemake features. We thought about the ability to monitor the run on the cloud (https://github.com/sequana/sequana/issues/414) but only on our cluster. In fact, a mechanism for a cloud run on a generic cluster would be a very nice feature indeed.
We are not familiar with Kubernetes but we could think of that in the future. Allowing for users to input a URL pointing to the config/snakefile files and having sequanix download them is also envisaged (https://github.com/sequana/sequana/issues/430).
