### 2017-08-10 sequana/sequana#437 @cokelaer
The plot [here](http://sequana.readthedocs.io/en/master/userguide.html#example-1-running-median-on-coverage) was sightly different from was you get when copying/pasting the code: the parameter 5001 in the code did not correspond to the picture (where 4001 was used). This is fixed but I am still wondering if that's what you meant by "I could not reproduce the plot".

I tried with the latest version of sequana on bioconda and using the standalone I got the same correct answer; same using the shell commands.

the notebook link appears in the "develop" version of the documentation. [here](http://sequana.readthedocs.io/en/develop/userguide.html#sequana-standalones)
