### 2017-08-10 sequana/sequana#436 @cokelaer
- we have created a requirements_dev loaded by setup.py and .travis
- we have updated requirements.txt loaded by setup.py and .travis
-  we have created a requirements_pipelines.txt loaded by .travis that contains all the pipeline requirements.
- environment.yml is unchanged and used only by readthedocs

even though this seems more completed since we have now 3 requirements*.txt they are now used through the documentation and factorised with the .travis and setup
