### 2017-08-10 #352 @bow
@cokelaer Thanks for the updates!

For the previous points:

1. Indeed, I realize splitting the current monolith will take more time. It's more of a development issue rather than usability for end users. Still it's something that needs consideration.

2. This is more on the usability end, but it's good to see you have considered (or are considering) it.

I'll be looking at the changes and responding to the issues later today ~ it seems be going along fine.
