### 2017-08-10 sequana/sequana#437 @bow
Aha, I've missed the notebook. It is indeed handy to have (especially with the interactive widget).

As for when I mentioned I could not reproduce, I meant I could not create the plots at all. I tried running them inside a notebook (the same environment as the sequana installation), but nothing was shown.

This was a verbatim copy-paste of the commands, though. I think it's the `%pylab inline` directive that needs to be added to my notebook.

I guess the example is also not meant to be run on its own Python script then?
