### 2017-08-10 sequana/sequana#441 @bow
Yes, this is more of a reproducibility issue for developers intending to play around with a specific release version. Right now, one has to manually trawl the commit messages. With a git tag, one can just checkout the tag and have the exact same state of the code as when it is released for that tag.
