### 2017-08-10 sequana/sequana#440 @bow
Almost there ~ the link to Travis still says `Travis.org` (missing `-ci`) and points to `travis.ci.org` (should be`travis-ci` instead of `travis.ci`) :).
