### 2017-08-11 sequana/sequana#437 @cokelaer
the example related to the enome coverage can be used inside the notebook (indeed with %pylab inline at the top of the notebook to show the images) or within a Python script. WE have wrap most of the code inside the standalone application sequana_coverage.
