### 2017-08-12 sequana/sequana#436 @bow
Ok, this looks more unified now. I tested it and could get a local dev setup with `pip install -e .` on a cloned repo (tried this in the `develop` branch on a freshly created virtualenv). It needs an additional `pip install PyQt5` for running sequanix under this setup but this is easily done.

It would be useful to also add some setup notes to the docs, but otherwise feel free to mark this issue as solved.
