### 2017-08-12 sequana/sequana#437 @bow
Ok ~ feel free to mark the issue closed. I tried it again (copying verbatim) the notebook you linked and it works ok. Replacing the BED file with the BED file in the current documentation also works.
