### 2017-08-12 sequana/sequana#440 @bow
Almost there ~ could you replace the displayed `Travis.org` link to either showing `travis-ci.org` or just `Travis CI`? It does already link to the correct URL, but the displayed text `Travis.org` points to something completely different.
