### 2017-08-12 sequana/sequana#436 @cokelaer
great. I uploaded a new version, tried with pip install sequana + pip install PyQt5 and it works indeed.
I close the issue.
