### 2017-08-12 sequana/sequana#441 @cokelaer
done for 0.4.2
