### 2017-08-14 sequana/sequana#441 @bow
Great, just #439 remaining then? Otherwise, I think this is ready :).
