### 2017-08-14 sequana/sequana#439 @cokelaer
@bow follow up on the RNAseq and atac-seq pipelines.

Following the tutorial and small changes, it appears that one can use the RNA-seq. There are still some hard-coded path but are not used by default. As for the atac-seq it is in the library but not yet released officially: it does not appear in the tutorial or sequanix or pipeline section in principles. We will fix the remaining hard-coded paths in the future though for the RNA-seq. I have updated the tutorial
http://sequana.readthedocs.io/en/master/tutorial.html#rna-seq to emphasize the part that must be edited to make it working.

I suspect this would not be an issue for JOSS acceptation but if you think this is a blocking issue, please let us know, otherwise we can close this issue for now. Note, however, that I have open an independent issue to keep track of this concern in https://github.com/sequana/sequana/issues/452
