### 2017-08-15 sequana/sequana#439 @bow
Hi @cokelaer,

If these are default values that are meant to be overwritten, I think so long as there is enough explanation in the documentation it is acceptable, but not preferred. In this case, it's more about leaking out private details of an institution's cluster to the outside world and also user experience.

However, if these values can not be overwritten and requires user to manually change it in the source directly (not via sequana), I think it's better to remove them first.

If I may go on to the technical details, can these strings not be replaced by empty strings Does sequana / sequanix understand that some config fields have values that must be given by users?
