### 2017-08-28 sequana/sequana#439 @cokelaer
@bow sorry for the delay in solving this last issue.

So, we slightly updated the configuration file of the rnaseq pipeline.

For now we won't remove all of the hard-coded parameters. However, these parameters have to be changed by the users anyway. Therefore they do not prevent the use of the pipelines. For example, the current default config file should be sufficient to run the pipeline locally as demonstrated in the online tutorial on the readthedocs (http://sequana.readthedocs.io/en/master/tutorial.html#rna-seq) ; I could re-run the suggested analysis, which takes about 5-10 minutes. Of course I needed to edit the configuration file as explained in the tutorial but the only section to be changed corresponds to the organism used in the example (genome section).
