### 2017-08-28 #352 @cokelaer
@bow @pjotrp
I've just commented on the pending issue (hard-coded links). The changes have been made on the develop branch and should be updated on the master very soon (tomorrow most probably).
