### 2017-08-29 #352 @arfon
Thanks @pjotrp. @cokelaer - please take a look at this PR https://github.com/sequana/sequana/pull/454 for some required changes for acceptance. Once you've made these could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
