### 2017-08-29 sequana/sequana#439 @bow
@cokelaer No worries, ok then ~ so long as you're aware of this issue. It's always safest to remove any traces of local environments for these things.
