### 2017-08-29 #352 @cokelaer
@arfon I've release a stable version 0.5.0 on pypi + release on github + zenodo:
https://zenodo.org/record/853160#.WaVZ6Wewyh8
Not sure whether you need the "all versions" DOI or the latest only.
For the latest, it is 10.5281/zenodo.853160
