### 2017-08-29 #352 @arfon
@cokelaer - one final set of changes here please https://github.com/sequana/sequana/pull/455
