### 2017-08-30 #352 @cokelaer
merged sequana/sequana#455
