### 2017-08-30 #352 @cokelaer
Note also that the current version of sequana tagged for zenodois 0.5.0 (not 0.4.1 used when we started the joss submission)
