### 2017-08-30 #352 @cokelaer
@bow @pjotrp Brilliant, thanks a lot. And again thank you for the review.

I've just noticed on the link above that the doi2bib link does not return the bibtex entry but a 'not found' message. Maybe we just need to be patient for the link to be active ?
