### 2017-08-30 #352 @arfon
> I've just noticed on the link above that the doi2bib link does not return the bibtex entry but a 'not found' message. Maybe we just need to be patient for the link to be active ?

This generally takes a couple of hours to resolve.
